==> 03_data_prep_world_bank.ipynb <==
# Data Prepping World Bank 
This notebook loads the required variables directly from the World Bank Database and merges them to the existing dataset containing UCDP and FORGE data. The following variables are obtained from the World Bank: oil rents, forest coverage, percentage of rural population and Internet usage. The variables have missing values and need to be imputed. 

### Load libraries -------
import pandas as pd
import matplotlib.pyplot as plt
from matplotlib import rcParams
import numpy as np
import os
import wbdata  # <--- use World Bank library to load data

# Pandas parameters
#pd.set_option('display.max_colwidth', -1)
#pd.set_option('display.max_rows', 2000)
#pd.set_option('display.max_columns', 500)

### Define plot parameters ------
plt.rcParams["font.family"] = "garamond"
plt.rcParams.update({'font.size': 22})
rcParams.update({'figure.autolayout': True})

### Define out paths ------

# check if out path exists, if it does not exists add the path
if not os.path.exists('C:\\Users\\frank\\Desktop\\master_thesis_out'):
    os.makedirs('C:\\Users\\frank\\Desktop\\master_thesis_out')

# add subfolders for the out path
out_paths = {
    "desciptive_plots_outcome": 'C:\\Users\\frank\\Desktop\\master_thesis_out\desciptive_plots_outcome',
    "desciptive_tables_outcome": 'C:\\Users\\frank\\Desktop\\master_thesis_out\desciptive_tables_outcome',
    "data": 'C:\\Users\\frank\\Desktop\\master_thesis_out\data'
}

for key, val in out_paths.items():
    if not os.path.exists(val):
        os.makedirs(val)
        
out_paths["desciptive_tables_outcome"]

### Load data --------
ucdp_forge = pd.read_csv(os.path.join(out_paths["data"], "ucdp_forge.csv"))
ucdp_forge.head(4)
len(ucdp_forge)

### Check country codes to merge with world bank -----------
# World bank data has their own country codes which are slightly different than iso codes
# check country codes directly on the webpage for example https://data.worldbank.org/country/algeria
con_codes = ucdp_forge.groupby(["StateAbb",'cname']).size().reset_index().rename(columns={0:'count'})
con_codes.head(3)

### Copy existing country codes and then replace codes to match world bank manually --------------------------- 
ucdp_forge["ccode_wb"] = ucdp_forge["StateAbb"]

# A. 
ucdp_forge.loc[ucdp_forge['StateAbb'] == "ALG", 'ccode_wb'] = "DZA"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "ANG", 'ccode_wb'] = "AGO"

# B.
ucdp_forge.loc[ucdp_forge['StateAbb'] == "BNG", 'ccode_wb'] = "BGD"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "BOS", 'ccode_wb'] = "BIH"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "BUI", 'ccode_wb'] = "BDI"

# C.
ucdp_forge.loc[ucdp_forge['StateAbb'] == "CAM", 'ccode_wb'] = "KHM"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "CDI", 'ccode_wb'] = "CIV"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "CEN", 'ccode_wb'] = "CAF"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "CHA", 'ccode_wb'] = "TCD"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "CON", 'ccode_wb'] = "COG"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "CRO", 'ccode_wb'] = "HRV"

# D.
ucdp_forge.loc[ucdp_forge['StateAbb'] == "DRC", 'ccode_wb'] = "COD"

# G. 
ucdp_forge.loc[ucdp_forge['StateAbb'] == "GRG", 'ccode_wb'] = "GEO"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "GUA", 'ccode_wb'] = "GTM"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "GUI", 'ccode_wb'] = "GIN"

# H. 
ucdp_forge.loc[ucdp_forge['StateAbb'] == "HAI", 'ccode_wb'] = "HTI"

# I.
ucdp_forge.loc[ucdp_forge['StateAbb'] == "INS", 'ccode_wb'] = "IDN"

# L.
ucdp_forge.loc[ucdp_forge['StateAbb'] == "LES", 'ccode_wb'] = "LSO"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "LEB", 'ccode_wb'] = "LBN" 
ucdp_forge.loc[ucdp_forge['StateAbb'] == "LIB", 'ccode_wb'] = "LBY"

# M. 
ucdp_forge.loc[ucdp_forge['StateAbb'] == "MAA", 'ccode_wb'] = "MRT"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "MAC", 'ccode_wb'] = "MKD"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "MLD", 'ccode_wb'] = "MDA"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "MOR", 'ccode_wb'] = "MAR"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "MYA", 'ccode_wb'] = "MMR"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "MZM", 'ccode_wb'] = "MOZ"

# N. 
ucdp_forge.loc[ucdp_forge['StateAbb'] == "NEP", 'ccode_wb'] = "NPL"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "NIG", 'ccode_wb'] = "NGA"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "NIR", 'ccode_wb'] = "NER"

# P.
ucdp_forge.loc[ucdp_forge['StateAbb'] == "PAR", 'ccode_wb'] = "PRY"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "PHI", 'ccode_wb'] = "PHL"

# R.
ucdp_forge.loc[ucdp_forge['StateAbb'] == "ROM", 'ccode_wb'] = "ROU"

# S.
ucdp_forge.loc[ucdp_forge['StateAbb'] == "SAL", 'ccode_wb'] = "SLV"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "SIE", 'ccode_wb'] = "SLE"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "SPN", 'ccode_wb'] = "ESP"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "SRI", 'ccode_wb'] = "LKA"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "SUD", 'ccode_wb'] = "SDN"

# T. 
ucdp_forge.loc[ucdp_forge['StateAbb'] == "TAJ", 'ccode_wb'] = "TJK"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "THI", 'ccode_wb'] = "THA"
ucdp_forge.loc[ucdp_forge['StateAbb'] == "TRI", 'ccode_wb'] = "TTO"

# U. 
ucdp_forge.loc[ucdp_forge['StateAbb'] == "UKG", 'ccode_wb'] = "GBR"

# Y. 
ucdp_forge.loc[ucdp_forge['StateAbb'] == "YAR", 'ccode_wb'] = "YEM" 
ucdp_forge.loc[ucdp_forge['StateAbb'] == "YUG", 'ccode_wb'] = "SRB" ### Recode Yugoslavia as Serbia

### Search for indicators in wb data --------------------
wbdata.search_indicators("rents")  

### Get data for oil rents  --------
# Source on the Internet: https://data.worldbank.org/indicator/NY.GDP.PETR.RT.ZS
oil = pd.DataFrame.from_dict(wbdata.get_data("NY.GDP.PETR.RT.ZS"))
oil_s = oil[["countryiso3code", "date", "value"]]
oil_s = oil_s.rename(columns={'countryiso3code': 'iso3', 'date': 'year', 'value': 'oil'})
oil_s.head(3)

### Merge oil variable with existing dataset (how = left) ----------
oil_s['year']=oil_s['year'].astype(int)
ucdp_forge_oil = pd.merge(ucdp_forge, oil_s, how='left', left_on=['year', 'ccode_wb'], right_on=['year', 'iso3'])
ucdp_forge_oil.head(4)

### Check if merge was successful ---------
codes = ucdp_forge_oil.groupby(["ccode_wb", 'iso3']).size()
len(ucdp_forge["StateAbb"].unique())
len(codes)

### Delete merge variable ----
ucdp_forge_oil = ucdp_forge_oil.drop(['iso3'], 1)
len(ucdp_forge_oil)

### Search for indicators in wb data --------------------
wbdata.search_indicators("forest area")  

### There is an issue with the database, data can only be loaded when specifying the required countries -----
countries = list(ucdp_forge["ccode_wb"].unique())
len(countries)

### Get data for forest coverage  --------
# Source on the Internet: https://data.worldbank.org/indicator/AG.LND.FRST.ZS
forest_cov = pd.DataFrame.from_dict(wbdata.get_data("AG.LND.FRST.ZS", country= countries))
forest_cov_s = forest_cov[["countryiso3code", "date", "value"]]
forest_cov_s = forest_cov_s.rename(columns={'countryiso3code': 'iso3', 'date': 'year', 'value': 'forest_cov'})
forest_cov_s.head(3)

### Merge forest coverage variable with existing dataset (how = left) -------
forest_cov_s['year']=forest_cov_s['year'].astype(int)
ucdp_forge_oil_forestcov = pd.merge(ucdp_forge_oil, forest_cov_s, how='left', left_on=['year', 'ccode_wb'], right_on=['year', 'iso3'])
ucdp_forge_oil_forestcov.head(4)

### Check if merge was successful ---------
codes = ucdp_forge_oil_forestcov.groupby(["ccode_wb", 'iso3']).size()
len(ucdp_forge["StateAbb"].unique())
len(codes)

### Delete merge variable ----
ucdp_forge_oil_forestcov = ucdp_forge_oil_forestcov.drop(['iso3'], 1)
len(ucdp_forge_oil_forestcov)

### Search for indicators in wb data --------------------
wbdata.search_indicators("rural population")  

### Get data for rural population  --------
rural = pd.DataFrame.from_dict(wbdata.get_data("SP.RUR.TOTL.ZS"))
rural_s = rural[["countryiso3code", "date", "value"]]
rural_s = rural_s.rename(columns={'countryiso3code': 'iso3', 'date': 'year', 'value': 'rural'})
rural_s.head(3)

### Merge rural population variable with existing dataset (how = left) -----
# Source on the Internet: https://data.worldbank.org/indicator/SP.RUR.TOTL.ZS
rural_s['year']=rural_s['year'].astype(int)
ucdp_forge_oil_forestcov_rural = pd.merge(ucdp_forge_oil_forestcov, rural_s, how='left', left_on=['year', 'ccode_wb'], right_on=['year', 'iso3'])
ucdp_forge_oil_forestcov_rural.head(4)

### Check if merge was successful ---------
codes = ucdp_forge_oil_forestcov_rural.groupby(["ccode_wb", 'iso3']).size()
len(ucdp_forge["StateAbb"].unique())
len(codes)

### Delete merge variable ----
ucdp_forge_oil_forestcov_rural = ucdp_forge_oil_forestcov_rural.drop(['iso3'], 1)
len(ucdp_forge_oil_forestcov_rural)

### Search for indicators in wb data --------------------
wbdata.search_indicators("internet")  

### Get data for internet usage  --------
# Source on the Internet: https://data.worldbank.org/indicator/IT.NET.USER.ZS
internet_use = pd.DataFrame.from_dict(wbdata.get_data("IT.NET.USER.ZS"))
internet_use_s = internet_use[["countryiso3code", "date", "value"]]
internet_use_s = internet_use_s.rename(columns={'countryiso3code': 'iso3', 'date': 'year', 'value': 'internet_use'})
internet_use_s.head(3)

### Merge internet usage variable with existing dataset (how = left) -------
internet_use_s['year']=internet_use_s['year'].astype(int)
ucdp_forge_oil_forestcov_rural_internetuse = pd.merge(ucdp_forge_oil_forestcov_rural, 
                                            internet_use_s, how='left', left_on=['year', 'ccode_wb'], right_on=['year', 'iso3'])
ucdp_forge_oil_forestcov_rural_internetuse.head(4)

### Check if merge was successful ---------
codes = ucdp_forge_oil_forestcov_rural_internetuse.groupby(["ccode_wb", 'iso3']).size()
len(ucdp_forge["StateAbb"].unique())
len(codes)

### Delete merge variable ----
ucdp_forge_oil_forestcov_rural_internetuse = ucdp_forge_oil_forestcov_rural_internetuse.drop(['iso3'], 1)
len(ucdp_forge_oil_forestcov_rural_internetuse)

### Check missing values in added variables ----
missing = ucdp_forge_oil_forestcov_rural_internetuse[['year',
 'StateAbb', 'forest_cov', 'rural', 'oil', 'internet_use']]
null_data = missing[missing.isnull().any(axis=1)]
null_data.sort_values(by=["StateAbb", "year"]).head(3) # rural has no missing values

### Which variables are in dataset -------------
dat_var = []
for col in ucdp_forge_oil_forestcov_rural_internetuse.columns:
    dat_var.append(col)
dat_var

### Reset index -----------
ucdp_forge_oil_forestcov_rural_internetuse.reset_index(drop=False, inplace=False)
ucdp_forge_oil_forestcov_rural_internetuse.head(3)

## Save dataset -----------
print(len(ucdp_forge_oil_forestcov_rural_internetuse))
ucdp_forge_oil_forestcov_rural_internetuse.to_csv(os.path.join(out_paths["data"], "ucdp_forge_wb.csv"), index=False, sep=',')
