==> 04_data_prep_QoG.ipynb <==
# Data Prepping Quality of Government
This notebook obtains indicators from the Quality of Government Standard Datasets and adds them to the existing dataset containing variables from UCDP, FORGE and the World Bank. The following variables are obtained from the Quality of Government data: quality of government, military expenditure index and military personnel index. The variables have missing values and need to be imputed.

### Load libraries -------
import pandas as pd
import matplotlib.pyplot as plt
from matplotlib import rcParams
import numpy as np
import os

# Pandas parameters
#pd.set_option('display.max_colwidth', -1)
#pd.set_option('display.max_rows', 2000)
#pd.set_option('display.max_columns', 500)

### Define plot parameters ------
plt.rcParams["font.family"] = "garamond"
plt.rcParams.update({'font.size': 22})
rcParams.update({'figure.autolayout': True})

### Define out paths ------

# check if out path exists, if it does not exists add the path
if not os.path.exists('C:\\Users\\frank\\Desktop\\master_thesis_out'):
    os.makedirs('C:\\Users\\frank\\Desktop\\master_thesis_out')

# add subfolders for the out path
out_paths = {
    "desciptive_plots_outcome": 'C:\\Users\\frank\\Desktop\\master_thesis_out\desciptive_plots_outcome',
    "desciptive_tables_outcome": 'C:\\Users\\frank\\Desktop\\master_thesis_out\desciptive_tables_outcome',
    "data": 'C:\\Users\\frank\\Desktop\\master_thesis_out\data'
}

for key, val in out_paths.items():
    if not os.path.exists(val):
        os.makedirs(val)
        
out_paths["desciptive_tables_outcome"]

### Load data --------
ucdp_forge_wb = pd.read_csv(os.path.join(out_paths["data"], "ucdp_forge_wb.csv"))
ucdp_forge_wb.head(4)
len(ucdp_forge_wb)

### Check existing cow country codes ----
# check manually with home page: https://correlatesofwar.org/data-sets/cow-country-codes
con_codes = ucdp_forge_wb.groupby(["StateAbb", 'ccode_wb', 'cname', 'ccode']).size().reset_index().rename(columns={0:'count'})
con_codes.head(3)

### Inconsistency in existing cow country codes -----
ucdp_forge_wb[["year", 'cname', 'ccode']].loc[ucdp_forge_wb["ccode"]==678] # this should be Yemen 679
# 678 refers to Yemen Arab Republic

### Change cow code for Yemen manually ------
ucdp_forge_wb.loc[ucdp_forge_wb['ccode'] == 678, 'ccode'] = 679
ucdp_forge_wb[["year", 'cname', 'ccode']].loc[ucdp_forge_wb["ccode"]==679]

### Load QoG Standard Dataset version Jan21 ----------
# Codebook: https://www.qogdata.pol.gu.se/data/codebook_std_jan21.pdf

qog = pd.read_csv("https://www.qogdata.pol.gu.se/data/qog_std_ts_jan21.csv", low_memory=False)
qog.head()

### Check cow country codes in qog data ----
# check manually with home page: https://correlatesofwar.org/data-sets/cow-country-codes
qog_codes = qog.groupby(["cname", "ccodecow"]).size().reset_index().rename(columns={0:'count'})
qog_codes.head(3)

### Subset qog data to only include variables which are needed ----
qog_s = qog[["ccodecow", "year", "icrg_qog", "bicc_milexp", "bicc_milper"]]
qog_s = qog_s.rename(columns={'ccodecow': 'ccode_qog', 'year': 'year', 'icrg_qog': 'gov_qual', 
                              'bicc_milexp': 'milexp_in', 'bicc_milper': 'milper_in'})
qog_s.head(3)

### Merge qog variables with existing dataset (how = left) --------
ucdp_forge_wb_qog = pd.merge(ucdp_forge_wb, qog_s, how='left', left_on=['year', 'ccode'], right_on=['year', 'ccode_qog'])
ucdp_forge_wb_qog.head(3)

### Check if merge was successful ---------
codes = ucdp_forge_wb_qog.groupby(["ccode_qog", 'ccode']).size()
ucdp_forge_wb["ccode"].unique()
len(codes)

### Delete merge variable ----
ucdp_forge_wb_qog = ucdp_forge_wb_qog.drop(['ccode_qog'], 1)
len(ucdp_forge_wb_qog)

### Which variables are in data -------------
dat_var = []
for col in ucdp_forge_wb_qog.columns:
    dat_var.append(col)
dat_var

### Check missing values in added variables ----
missing = ucdp_forge_wb_qog[['year',
 'cname', 'ccode',
 'gov_qual',
 'milexp_in',
 'milper_in']]
null_data = missing[missing.isnull().any(axis=1)]
null_data.sort_values(by=["cname", 'ccode', "year"]).head(3)

### Check countries which are completely missing since this indicates a merging issue ----
qog[["ccodecow", "year", "icrg_qog", "bicc_milexp", "bicc_milper"]].loc[qog["ccodecow"]==346] # Bosnia and Herzegovina, check
qog[["ccodecow", "year", "icrg_qog", "bicc_milexp", "bicc_milper"]].loc[qog["ccodecow"]==530] # Ethiopia, check
qog[["ccodecow", "year", "icrg_qog", "bicc_milexp", "bicc_milper"]].loc[qog["ccodecow"]==625] # Sudan, check
qog[["ccodecow", "year", "icrg_qog", "bicc_milexp", "bicc_milper"]].loc[qog["ccodecow"]==365] # USSR, check
qog[["ccodecow", "year", "icrg_qog", "bicc_milexp", "bicc_milper"]].loc[qog["ccodecow"]==678] # Yemen <--- not in qog data, should be checked (see above)

#### Reset index -----------
ucdp_forge_wb_qog.reset_index(drop=False, inplace=False)
ucdp_forge_wb_qog.head(3)

### Save dataset -----------
ucdp_forge_wb_qog.to_csv(os.path.join(out_paths["data"], "ucdp_forge_wb_qog.csv"), index=False, sep=',')
len(ucdp_forge_wb_qog)
