==> 05_data_prep_polity.ipynb <==
# Data Prepping Polity5
This notebook loads data from the polity5 project and merges them with the existing dataset, including data from UCDP, FORGE, World Bank and the Quality of Government Dataset. The following variable is added: the combined polity score. The variable has missing values which need to be imputed.

### Load libraries -------
import pandas as pd
import matplotlib.pyplot as plt
from matplotlib import rcParams
import numpy as np
import os

# Pandas parameters
#pd.set_option('display.max_colwidth', -1)
#pd.set_option('display.max_rows', 2000)
#pd.set_option('display.max_columns', 500)

### Define plot parameters ------
plt.rcParams["font.family"] = "garamond"
plt.rcParams.update({'font.size': 22})
rcParams.update({'figure.autolayout': True})

### Define out paths ------

# check if out path exists, if it does not exists add the path
if not os.path.exists('C:\\Users\\frank\\Desktop\\master_thesis_out'):
    os.makedirs('C:\\Users\\frank\\Desktop\\master_thesis_out')

# add subfolders for the out path
out_paths = {
    "desciptive_plots_outcome": 'C:\\Users\\frank\\Desktop\\master_thesis_out\desciptive_plots_outcome',
    "desciptive_tables_outcome": 'C:\\Users\\frank\\Desktop\\master_thesis_out\desciptive_tables_outcome',
    "data": 'C:\\Users\\frank\\Desktop\\master_thesis_out\data'
}

for key, val in out_paths.items():
    if not os.path.exists(val):
        os.makedirs(val)
        
out_paths["desciptive_tables_outcome"]

### Load data --------
ucdp_forge_wb_qog = pd.read_csv(os.path.join(out_paths["data"], "ucdp_forge_wb_qog.csv"))
ucdp_forge_wb_qog.head(4)
len(ucdp_forge_wb_qog)

### Load Polity5: Regime Authority Characteristics and Transitions Datasets ----------
# Codebook: http://www.systemicpeace.org/inscr/p5manualv2018.pdf

polity = pd.read_excel("https://www.systemicpeace.org/inscr/p5v2018.xls")
polity.head()

### Check cow country codes in polity data ----
# check manually with home page: https://correlatesofwar.org/data-sets/cow-country-codes
polity_codes = polity.groupby(["country", "ccode"]).size().reset_index().rename(columns={0:'count'})
polity_codes.head(3)

### Make subset to only included the needed variables -----
polity_s = polity[["ccode", "year", "polity2"]]
polity_s = polity_s.rename(columns={'ccode': 'ccode_polity', 'year': 'year', 'polity2': 'polity2'})
polity_s.head(3)

### Merge polity variable with existing dataset (how = left) -----
ucdp_forge_wb_qog_polity = pd.merge(ucdp_forge_wb_qog, polity_s, how='left',
                                                   left_on=['year', 'ccode'], right_on=['year', 'ccode_polity'])
ucdp_forge_wb_qog_polity.head(3)

### Check if merge was successful ---------
codes = ucdp_forge_wb_qog_polity.groupby(["ccode_polity", 'ccode']).size()
ucdp_forge_wb_qog["ccode"].unique()
len(codes)

### Delete merge variable ----
ucdp_forge_wb_qog_polity = ucdp_forge_wb_qog_polity.drop(['ccode_polity'], 1)
len(ucdp_forge_wb_qog_polity)

### Which variables are in data -------------
dat_var = []
for col in ucdp_forge_wb_qog_polity.columns:
    dat_var.append(col)
dat_var

### Check missing values in added variables ----
missing = ucdp_forge_wb_qog_polity[['year',
 'cname', 'ccode', 'polity2']]
null_data = missing[missing.isnull().any(axis=1)]
null_data.sort_values(by=["cname", "year"]).head(3)

### Check countries which missing values to avoid merging issue ----
polity[["country", "ccode", 'polity2', 'year']].loc[polity["ccode"]==700] # Afghanistan, check
polity[["country", "ccode", 'polity2', 'year']].loc[polity["ccode"]==530] # Ethiopia, check
polity[["country", "ccode", 'polity2', 'year']].loc[polity["ccode"]==645] # Iraq, check
polity[["country", "ccode", 'polity2', 'year']].loc[polity["ccode"]==365].head(3) # USSR, check

### Reset index -----------
ucdp_forge_wb_qog_polity.reset_index(drop=False, inplace=False)
ucdp_forge_wb_qog_polity.head(3)

### Save dataset -----------
ucdp_forge_wb_qog_polity.to_csv(os.path.join(out_paths["data"], 
                                              "ucdp_forge_wb_qog_polity.csv"), 
                                               index=False, sep=',')
len(ucdp_forge_wb_qog_polity)
